Test SSR.
