+++ packages/style/src/common/syntax.ts
@@ -65,7 +65,10 @@ export function formatRule(
 ): CSS {
   let rule = `.${className}${selector} { ${block} }`;
 
-  if (conditions) {
+  // Server-side rendering recursively creates CSS rules to collapse
+  // conditionals to their smallest representation, so we need to avoid
+  // wrapping with the outer conditional for this to work correctly.
+  if (conditions && !process.env.AESTHETIC_SSR) {
     arrayLoop(
       conditions,
       (condition) => {
